requirements.txt
Pillow
numpy
torch
diffusers
